+++ res/views/components/form.blade.php
@@ -1,7 +1,5 @@
-@tag('form', $attributes->merge(['method' => $method === 'GET' ? $method : 'POST', 'action' => $action, 'enctype' => $hasFiles ? 'multipart/form-data' : false]))
+@tag('form', $attributes->merge([...compact('method', 'action'), 'enctype' => $hasFiles ? 'multipart/form-data' : false]))
     @csrf
-    @if ($method !== 'GET' && $method !== 'POST')
-        @method($method)
-    @endif
+    @method($method)
     {{ $slot }}
 @endtag