+++ res/views/components/link.blade.php
@@ -1,7 +1,3 @@
-@php
-$href = isset($route) ? route($route, $params) : url($href);
-@endphp
-
 @tag('a', $attributes->merge(compact('href')))
     <s-text :value="$text">
         {{ $slot }}